[c] - secure routes with jwt verify
