+++ backend/middlewares/serviceTokenVerify.ts
@@ -1,22 +1,31 @@
-import { NextFunction, Request, Response } from 'express'
-import AppResponse from '../misc/AppResponse.ts'
+import { Request, Response, NextFunction } from 'express';
+import jwt from 'jsonwebtoken';
+import AppResponse from '../misc/AppResponse.ts';
 
-const serviceTokenVerify = (
-  req: Request,
-  res: Response,
-  next: NextFunction,
-) => {
-  const authHeader = req.headers.authorization
+const secretKey = Deno.env.get('JWT_SECRET');
+
+const serviceTokenVerify = (req: Request, res: Response, next: NextFunction) => {
+  const authHeader = req.headers.authorization;
   if (!authHeader || !authHeader.startsWith('Bearer ')) {
-    return AppResponse.unauthorized(res)
+    return AppResponse.unauthorized(res);
   }
 
-  const token = authHeader.split(' ')[1]
-  if (token !== Deno.env.get('SUPABASE_SERVICE_ROLE_KEY')) {
-    return AppResponse.unauthorized(res)
+  const token = authHeader.split(' ')[1];
+
+
+  if (!secretKey) {
+    console.error('JWT secret key is not defined in the environment variables.');
+    return AppResponse.internalServerError(res);
   }
 
-  next()
-}
+  jwt.verify(token, secretKey, (err, decoded) => {
+    if (err) {
+      return AppResponse.unauthorized(res);
+    }
+    // Optional: Attach user or decoded token to request object
+    // req.user = decoded;
+    next();
+  });
+};
 
-export default serviceTokenVerify
+export default serviceTokenVerify;
