+++ backend/routes/Api.ts
@@ -2,6 +2,7 @@ import { Router } from 'express'
 
 import Paths from '../constants/Paths.ts'
 import broadcastController from '../controllers/BroadcastController.ts'
+import serviceRoleKeyVerify from '../middlewares/serviceRoleKeyVerify.ts'
 import serviceTokenVerify from '../middlewares/serviceTokenVerify.ts'
 
 const apiRouter = Router()
@@ -10,32 +11,37 @@ const broadcastRouter = Router()
 
 broadcastRouter.get(
   Paths.Broadcast.Make,
-  serviceTokenVerify,
+  serviceRoleKeyVerify,
   broadcastController.makeBroadcast,
 )
 
 broadcastRouter.get(
   Paths.Broadcast.SendNow,
+  serviceTokenVerify,
   broadcastController.sendNow,
 )
 
 broadcastRouter.get(
   Paths.Broadcast.Draft,
+  serviceTokenVerify,
   broadcastController.sendDraft,
 )
 
 broadcastRouter.get(
   Paths.Broadcast.All,
+  serviceTokenVerify,
   broadcastController.getAll,
 )
 
 broadcastRouter.patch(
   Paths.Broadcast.ID,
+  serviceTokenVerify,
   broadcastController.patch,
 )
 
 broadcastRouter.get(
   Paths.Broadcast.UpdateTwilioStatus,
+  serviceRoleKeyVerify,
   broadcastController.updateTwilioStatus,
 )
