+++ package.json
@@ -5,14 +5,17 @@
   "devDependencies": {
     "@types/express": "^4.17.21",
     "@types/morgan": "^1.9.9",
-    "postgres": "^3.4.3",
-    "node-mocks-http": "^1.14.1",
     "cors": "^2.8.5",
     "drizzle-orm": "^0.29.3",
     "express": "^4.18.2",
     "express-async-errors": "^3.1.1",
     "express-validator": "^7.0.1",
     "helmet": "^7.1.0",
-    "morgan": "^1.10.0"
+    "morgan": "^1.10.0",
+    "node-mocks-http": "^1.14.1",
+    "postgres": "^3.4.3"
+  },
+  "dependencies": {
+    "jsonwebtoken": "^9.0.2"
   }
 }
